cleanup(storage)!: consistent types for generation
Both `generation` and `metageneration` are `Int64` in the generated
types. It is easier to reason and work with these things if we use the
same types (as opposed to `UInt64`) in our own types.

+++ Tests/StorageW1R3/StorageOperations.swift
@@ -63,7 +63,7 @@ enum StorageOperations {
     client: StorageClient, object: GoogleCloudStorage.Object, crc32cEnabled: Bool
   ) async -> (transferSize: Int, error: (any Error)?) {
     let options = ReadObjectOptions().with {
-      $0.generation = if object.generation > 0 { UInt64(object.generation) } else { nil }
+      $0.generation = if object.generation > 0 { object.generation } else { nil }
       $0.checksums = crc32cEnabled ? .default : .none
     }
